push chage

+++ backend/src/main/java/com/project/backend/ResponseDto/CartPricingResponseDto.java
@@ -1,5 +1,6 @@
 package com.project.backend.ResponseDto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.Builder;
@@ -10,15 +11,15 @@ import lombok.Data;
 public class CartPricingResponseDto {
 
 	 private List<CartItemResponseDto> items;
-	    private Double subtotal;
-	    private Double taxAmount;
-	    private Double shippingCharges;
-	    private Double discountAmount;
-	    private Double finalAmount;
+	    private BigDecimal subtotal;
+	    private BigDecimal taxAmount;
+	    private BigDecimal shippingCharges;
+	    private BigDecimal discountAmount;
+	    private BigDecimal finalAmount;
 	    private String appliedCoupon;
 	    private Boolean couponApplied;
 	    private String message;
 	    private Integer totalItems;
-	    private Double totalMrp;
-	    private Double totalSavings;
+	    private BigDecimal totalMrp;
+ private BigDecimal totalSavings;
 }
